+++ POO/POO3/TesteXUnit/ContaEspecialTest.cs
@@ -15,7 +15,7 @@ namespace TesteXUnit
         [InlineData(1000, 530, 600)]
         [InlineData(1000, 500, 700)]
         [InlineData(1000, 1010, 1500)]
-        public void DeveSacarDinheiroCasoTenhaSaldoEDescontarATaxa(double limite, double saldo, double saque){
+        public void DeveSacarDinheiroCasoTenhaSaldoEDescontarATaxa(decimal limite, decimal saldo, decimal saque){
             var conta = new ContaEspecial(1234, limite);
             conta.Saldo = saldo;
             Assert.Equal(saque, conta.Sacar(saque));
@@ -25,7 +25,7 @@ namespace TesteXUnit
         [InlineData(1000, -350, 500, 150)]
         [InlineData(650, 100, 490, 590)]
         [InlineData(100, 1000,  995, 1995)]
-        public void DeveDepositar(double limite, double saldo, double deposito, double valorEsperado){
+        public void DeveDepositar(decimal limite, decimal saldo, decimal deposito, decimal valorEsperado){
             var conta = new ContaEspecial(1234, limite);
             conta.Saldo = saldo;
             conta.Depositar(deposito);
@@ -37,7 +37,7 @@ namespace TesteXUnit
         [InlineData(1000, -350,  5150, 0)]
         [InlineData(650,  100,  5000, 0)]
         [InlineData(100,  1000, 10000, 0)]
-        public void NaoDeveSacarContaNaoTemSaldoMesmoComLimite(double limite, double saldo, double saque, double valorEsperado){
+        public void NaoDeveSacarContaNaoTemSaldoMesmoComLimite(decimal limite, decimal saldo, decimal saque, decimal valorEsperado){
             var conta = new ContaEspecial(1234, limite);
             conta.Saldo = saldo;
             Assert.Equal(valorEsperado, conta.Sacar(saque));
