+++ Questao4/Entidade/Aluno.cs
@@ -3,6 +3,6 @@ namespace Questao4.Entidade
     public class Aluno
     {
         public string Nome { get; set; }
-        public decimal Nota { get; set; }
+        public double Nota { get; set; }
     }
 }
